fix warning C4267: '=': conversion from 'size_t' to '_Ty', possible loss of data

+++ include/cppjieba/HMMSegment.hpp
@@ -159,7 +159,7 @@ class HMMSegment: public SegmentBase {
           tmp = weight[old] + model_->transProb[preY][y] + emitProb;
           if (tmp > weight[now]) {
             weight[now] = tmp;
-            path[now] = preY;
+            path[now] = static_cast<int>(preY);
           }
         }
       }
